requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.7
beautifulsoup4==4.8.2
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
confusable-homoglyphs==3.2.0
Django==2.2.9
django-admin-interface==0.12.0
django-appconf==1.0.3
django-apptemplates==1.4
django-celery==3.3.1
django-ckeditor==5.6.1
django-colorfield==0.1.15
django-compressor==2.4
django-countries==5.5
django-crispy-forms==1.7.2
django-db-signals==0.1.1
django-fieldsets-with-inlines==0.5
django-filter==2.0.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-google-analytics-app==4.4.0
django-hitcount==1.3.1
django-ipware==2.1.0
django-jet==1.0.8
django-js-asset==1.1.0
django-libsass==0.8
django-mathfilters==0.4.0
django-registration==3.0
django-sass-processor==0.8
django-tracking-analyzer==1.0.1
django-user-agents==0.4.0
django-wysiwyg==0.8.0
feedparser==5.2.1
geoip2==3.0.0
google-api-python-client==1.4.1
httplib2==0.15.0
idna==2.8
kombu==3.0.37
libsass==0.19.4
maxminddb==1.5.2
mccabe==0.6.1
oauth2client==4.1.3
Pillow==7.0.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pytz==2018.7
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
structlog==19.2.0
ua-parser==0.8.0
uritemplate==3.0.1
urllib3==1.25.7
user-agents==2.0
